move scaling code
