+++ src/openms/source/FILTERING/SCALING/Normalizer.cpp
@@ -7,7 +7,7 @@
 // --------------------------------------------------------------------------
 //
 
-#include <OpenMS/FILTERING/TRANSFORMERS/Normalizer.h>
+#include <OpenMS/FILTERING/SCALING/Normalizer.h>
 
 using namespace std;
 namespace OpenMS
