+++ src/tests/class_tests/openms/source/RankScaler_test.cpp
@@ -12,7 +12,7 @@
 
 ///////////////////////////
 
-#include <OpenMS/FILTERING/TRANSFORMERS/RankScaler.h>
+#include <OpenMS/FILTERING/SCALING/RankScaler.h>
 #include <OpenMS/KERNEL/StandardTypes.h>
 #include <OpenMS/FORMAT/DTAFile.h>
