+++ src/tests/class_tests/openms/source/SpectrumAlignmentScore_test.cpp
@@ -14,7 +14,7 @@
 #include <iostream>
 
 #include <OpenMS/COMPARISON/SpectrumAlignmentScore.h>
-#include <OpenMS/FILTERING/TRANSFORMERS/Normalizer.h>
+#include <OpenMS/FILTERING/SCALING/Normalizer.h>
 #include <OpenMS/FORMAT/DTAFile.h>
 
 ///////////////////////////
