+++ src/tests/class_tests/openms/source/ZhangSimilarityScore_test.cpp
@@ -15,7 +15,7 @@
 
 #include <OpenMS/COMPARISON/ZhangSimilarityScore.h>
 #include <OpenMS/FORMAT/DTAFile.h>
-#include <OpenMS/FILTERING/TRANSFORMERS/Normalizer.h>
+#include <OpenMS/FILTERING/SCALING/Normalizer.h>
 
 ///////////////////////////
