+++ src/topp/NucleicAcidSearchEngine.cpp
@@ -42,7 +42,7 @@
 #include <OpenMS/FILTERING/TRANSFORMERS/ThresholdMower.h>
 #include <OpenMS/FILTERING/TRANSFORMERS/NLargest.h>
 #include <OpenMS/FILTERING/TRANSFORMERS/WindowMower.h>
-#include <OpenMS/FILTERING/TRANSFORMERS/Normalizer.h>
+#include <OpenMS/FILTERING/SCALING/Normalizer.h>
 
 // spectra comparison
 #include <OpenMS/CHEMISTRY/NucleicAcidSpectrumGenerator.h>
